fix: forward and reverse aunction

+++ src/api/store/auctions/reverse/[id]/route.ts
@@ -24,12 +24,24 @@ export async function POST(
 ): Promise<void> {
   const id = req.params.id;
 
+  const customer_id = req.user.customer_id
+  if( customer_id ) {
+ 
+
   const auctionService = req.scope.resolve("auctionService") as AuctionService;
 
+  const customerAuction =  await auctionService.retrieve(id);
+
+  if(customerAuction.created_by == customer_id)
+  {
   const auction = await auctionService.update(id, req.body);
 
   res.status(200).json({ auction });
   return;
+  }
+}
+res.sendStatus(401)
+
 }
 
 export async function DELETE(
@@ -38,10 +50,20 @@ export async function DELETE(
 ): Promise<void> {
   const id = req.params.id;
 
+  const customer_id = req.user.customer_id
+  if( customer_id ) {
+ 
   const auctionService = req.scope.resolve("auctionService") as AuctionService;
 
+  const customerAuction =  await auctionService.retrieve(id);
+
+  if(customerAuction.created_by == customer_id)
+  {
   await auctionService.delete(id);
 
   res.status(200).json({});
   return;
+  }
+}
+  res.sendStatus(401)
 }
